Resolved one conflict in LASSIE/src/main.cpp: both branches added a
new offscreen-test block to the same DISSCO_ENABLE_UI_LAYOUT_TESTS
chain (main's DISSCO_TEST_WINDOWS_FONT/DISSCO_TEST_APPLICATION_IDENTITY
checks, and this branch's DISSCO_TEST_MODIFIER_ADD_REFRESH) -- kept
both, since neither depends on the other.

+++ CMOD/src/Define.h
@@ -70,8 +70,8 @@ static const double WELL_TEMP_INCR = pow(2, 1./12.);
 //static const double WELL_TEMP_INCR = pow(2, 1./24.);
 //static const double WELL_TEMP_INCR = 1.0594631;
 // static const float MAX_SONES = 256.;
-static const float FIRST_CONST = -5.54;
-static const float SECOND_CONST = -1.84;
+static const float FIRST_CONST = static_cast<float>(-5.54);
+static const float SECOND_CONST = static_cast<float>(-1.84);
 static const int SAMPLING_RATE = 44100;
 
 // constants for the Event class
